Create CLstatistics service

+++ Projet_POO_Groupe6/MyForm.h
@@ -3495,7 +3495,7 @@ namespace ProjetPOOGroupe6 {
 		this->dgv_emp->DataMember = "Rsl";
 	}
 	private: System::Void addEmp_Click(System::Object^ sender, System::EventArgs^ e) {
-		int streetNumber = System::Convert::ToInt64(this->text_streetNumber->Text);
+		int streetNumber = System::Convert::ToInt32(this->text_streetNumber->Text);
 		this->dgv_emp->Refresh();
 ////////
 /*Attention le hire date doit être pris automatiquement dans la requête sql avec un GETDATE() car plus aucun champs ne permet de le rentrer*/
@@ -3513,8 +3513,8 @@ namespace ProjetPOOGroupe6 {
 		this->dgv_emp->DataMember = "Rsl";
 	}
 	private: System::Void changeEmp_Click(System::Object^ sender, System::EventArgs^ e) {
-		int streetNumber = System::Convert::ToInt64(this->StreetNumChangeEmp->Text);
-		int employeeID = System::Convert::ToInt64(this->text_EmpIdChangeEmp->Text);
+		int streetNumber = System::Convert::ToInt32(this->StreetNumChangeEmp->Text);
+		int employeeID = System::Convert::ToInt32(this->text_EmpIdChangeEmp->Text);
 		this->dgv_emp->Refresh();
 		this->oSVCemployees->updateEmp(employeeID, this->text_chiefIdChangeEmp->Text, this->text_fistNameChangeEmp->Text, this->text_empLastName->Text, (this->hireChangeEmp_datePicker->Value).ToString("yyyy-MM-dd"), streetNumber, this->text_streetNameChangeEmp->Text, this->text_cityNameChangeEmp->Text, this->text_ZipCodeChangeEmp->Text);
 		this->oDs_employees = this->oSVCemployees->displayEmp("Rsl");
